# LabWork_45/docs/swagger.yaml
basePath: /
definitions:
  model.Login:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  model.SignUp:
    properties:
      login:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  model.User:
    properties:
      id:
        type: string
      login:
        type: string
      password:
        items:
          type: integer
        type: array
      refresh_token:
        items:
          type: integer
        type: array
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Diploma Documentation.
  title: Lab
  version: "1.0"
paths:
  /auth/delete:
    delete:
      description: Delete a user from the database
      produces:
      - application/json
      responses:
        "200":
          description: All users in system
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete
      tags:
      - User methods
  /auth/getall:
    get:
      description: GetAll function returns list of all users in database(test function)
      produces:
      - application/json
      responses:
        "200":
          description: All users in system
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "404":
          description: Error message
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: GetAll
      tags:
      - temporary methods
  /auth/getpersonalinfo:
    get:
      description: Fetch personal information of the active user based on the access
        token provided in the Authorization header.
      produces:
      - application/json
      responses:
        "200":
          description: User's personal information
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Personal Info
      tags:
      - User methods
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user and returns access and refresh tokens
      parameters:
      - description: Login details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Login'
      produces:
      - application/json
      responses:
        "200":
          description: ' Generating access and refresh tokens'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Error message
          schema:
            type: string
      summary: Login
      tags:
      - Authentication methods
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Creates a new user in the system
      parameters:
      - description: Sign up details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.SignUp'
      produces:
      - application/json
      responses:
        "200":
          description: User has been registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: SignUp
      tags:
      - Authentication methods
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"